implement flex <*> wildcard start condition
flex offers * as a shorthand to mean all available conditions.
it's not entirely clear how it interacts with a mix of inclusive
and exclusive start conditions, but from my reading of the flex
manual, chapter 8, it appears that when at least one exclusive
scon is in the list, exclusive behaviour is expected, which is
confirmed by simple testcases where lex and flex behaviour was
compared.
analysis of some other lex files, including testcases from flex
itself suggests that usually only either inclusive or exclusive
scons are used, but not mixed.

a heavy user of this feature is linux' dtc (device tree configuration)
program, which after this change can now successfully be compiled
without flex installed.