wcio.c: include ldefs.h instead of using a wrong error() prototype
the real prototype uses const char*, not char*.
instead of fixing the duplicate declaration, we rather fix the
ldefs.h compatibility issues with wcio.c:
ldefs.h uses a hackish redefinition of putwc/getwc to point
to the version in wcio.c.
we need to undef these 2 macros in order to prevent endless
recursion.