main.c: implement WHOLE_PROGRAM mode
when this preprocessor define is set, main.c includes all other
translation units. this way it can be used with -fwhole-program
for minimal binary size, in a similar way than the well known
"amalgamation", but without sucking at it, because the files
can still be built efficiently in parallel, if desired.

usage: cc -DWHOLE_PROGRAM main.c -o lex